pyproject.toml
[project]
name = "guestbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-shell-plus>=1.1.7",
    "djangorestframework>=3.16.1",
    "pre-commit>=4.3.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
]
